+++ test/async_extension_test.dart
@@ -553,6 +553,26 @@ void main() {
     });
   });
 
+  group('FutureNullableExtension', () {
+    test('Iterable', () async {
+      Future<int?> futureInt1 = Future.value(100);
+      expect(await futureInt1.orElseAsync(101), equals(100));
+
+      Future<int?> futureInt2 = Future.value(null);
+      expect(await futureInt2.orElseAsync(101), equals(101));
+    });
+  });
+
+  group('FutureOrNullableExtension', () {
+    test('Iterable', () {
+      FutureOr<int?> futureOrInt1 = 100;
+      expect(futureOrInt1.orElseAsync(101), equals(100));
+
+      FutureOr<int?> futureOrInt2;
+      expect(futureOrInt2.orElseAsync(101), equals(101));
+    });
+  });
+
   group('FutureOrIterableExtension', () {
     test('Iterable', () {
       FutureOr<Iterable<int>> futureOrItr = [1, 2, 3];